added javadoc to vbybetc

+++ src/test/java/com/softserve/edu/oms/tests/vbybetc/TC56vbTest.java
@@ -12,6 +12,9 @@ import org.testng.Assert;
 import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
 
+/**
+ * <h1>Verify that Login field is case insensitive</h1>
+ */
 public class TC56vbTest extends TestRunner{
 
     @DataProvider
@@ -28,6 +31,10 @@ public class TC56vbTest extends TestRunner{
         };
     }
 
+    /**
+     * just for login
+     * @param admUser
+     */
     @Test(dataProvider = "admUser")
     public void PreconditionTest(IUser admUser) {
 
@@ -35,6 +42,25 @@ public class TC56vbTest extends TestRunner{
                 .clickAdministrationTab()
                 .goToCreateNewUserPage();
     }
+    /**
+     * This test goes to Create New User Page,
+     * 1) executes SQL query to verify that nonExistingUser from UserRepository
+     * really is not in DB
+     * 2) If he, for some reason, exists - then generates new random Login
+     * 3) Then the data for nonExistingUser are set in the form "Create new user"
+     * 4) the fields "Role" and "Region" are not used, and leave as default
+     * 5) the same Login, but in reverse case is set in Login field
+     * 6) error message should appear - if so test passed
+     *
+     * <p>Note: there are two data providers and thus two tests, because
+     * first we login as Admin and then trying to create New User</p>
+     *
+     * @author Viktoriia Bybel
+     * @version 1.0
+     * @since 15.12.16
+     * @param nonExistingUser
+     * @see UserRepository
+     */
 
     @Test(dataProvider = "nonExistingUser")
     public void UniqueUserCreatingTest(IUser nonExistingUser)  {
@@ -57,13 +83,16 @@ public class TC56vbTest extends TestRunner{
                 .setEmailInput(nonExistingUser.getEmail())
                 .clickCreateButton();
 
-        CreateNewUserPage newUserPageAgain = new AdministrationPage(driver)
-                .goToCreateNewUserPage().setLoginInput(nonExistingLogin.toUpperCase());
+        if (!(dbUtils.verifyThatUserIsInDB(nonExistingLogin))) {
 
-        Assert.assertTrue(newUserPageAgain.getLoginErrorMessageText().contains("already in use"));
+            CreateNewUserPage newUserPageAgain = new AdministrationPage(driver)
+                    .goToCreateNewUserPage().setLoginInput(nonExistingLogin.toUpperCase());
 
-        dbUtils.deleteUsersFromDB(SQLQueries.DELETE_FROM_USERS_WHERE_LOGIN_EQUALS.getQuery(),
-                nonExistingUser.getLoginname());
+            Assert.assertTrue(newUserPageAgain.getLoginErrorMessageText().contains("already in use"));
+
+            dbUtils.deleteUsersFromDB(SQLQueries.DELETE_FROM_USERS_WHERE_LOGIN_EQUALS.getQuery(),
+                    nonExistingUser.getLoginname());
+        }
     }
 
 }
